finished implementing new brush action. Also fixed hardness in eraser tool.
